bill-extractor/requirements.txt
pytesseract
pdfplumber
pillow
opencv-python
streamlit
